Bug 1228314 - added static_cast<int64>  in order to avoid overflow. r=seth
MozReview-Commit-ID: KY3qpIlzE5K

UltraBlame original commit: 9727cdebb2eeb25a1d9452b7fa5e0583e879d0af

+++ image/SurfaceCache.cpp
@@ -327,7 +327,8 @@ public:
       
       
       
-      int64_t idealArea = idealKey.Size().width * idealKey.Size().height;
+      int64_t idealArea = static_cast<int64_t>(idealKey.Size().width) *
+        static_cast<int64_t>(idealKey.Size().height);
       int64_t surfaceArea = aSurfaceKey.Size().width * aSurfaceKey.Size().height;
       int64_t bestMatchArea =
         bestMatchKey.Size().width * bestMatchKey.Size().height;
